move confirmation UI code out of tx runner and into the caller
